Folder structure, lots of polish
- Polished a load of XYZ tracking positional mistakes
- Y is now available for setting destination. Very important when trying
to move upwards and downwards (currently only the ramp, but it is
definitely going to be important on the map)
- Flag is way easier to pick up now
- Reorganized folder structure
- Set the player into the ignore raycast layer, this was extremely
important to make sure if you click on the same spot before and after
reaching the destination that the coordinates are always the same.
Before the raycast reflected from the player and it made the movement
extremely clunky. This one got me stuck for over an hour.