Map Design implementation, team sides, object interaction
- Implemented the map elements to mostly match the sketch I've been
provided with
- Flag can now be assigned to a team, capture works properly now on the
entire surface of its floor instead of only a single point
- Related to the point above, interactable objects can now have 2
interaction states. One of them is a single reach point, recommended for
objects you can't walk on, like the flag. The second one is the entire
area it covers from the top perspective (therefore no account for Y
coord), example is the Flag floor
- Adjusted lots of mathematical mistakes, the code is more efficient now
- Fiddled with the player speed a bit. It's very fast for now just to
make testing easier, it can be set in the Player object's NavMeshAgent
component along with other stuff concerning the player navigation
basically